requirements.txt
dhanhq>=2.0.0
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0
colorama>=0.4.6
tabulate>=0.9.0
